+++ tests/test_doctests.py
@@ -11,16 +11,17 @@ import pytest
 
 # Modules that have working doctests
 DOCTEST_MODULES = [
-    "torchsparsegradutils.indexed_matmul",
-    "torchsparsegradutils.sparse_logsumexp",
-    "torchsparsegradutils.sparse_lstsq",
-    "torchsparsegradutils.sparse_matmul",
-    "torchsparsegradutils.sparse_solve",
-    "torchsparsegradutils.utils.bicgstab",
-    "torchsparsegradutils.utils.linear_cg",
-    "torchsparsegradutils.utils.minres",
+    "torchsparsegradutils.ops.indexed_matmul",
+    "torchsparsegradutils.ops.logsumexp",
+    "torchsparsegradutils.ops.lstsq",
+    "torchsparsegradutils.ops.matmul",
+    "torchsparsegradutils.ops.triangular_solve",
+    "torchsparsegradutils.ops.generic_solve",
+    "torchsparsegradutils.solvers.bicgstab",
+    "torchsparsegradutils.solvers.cg",
+    "torchsparsegradutils.solvers.minres",
     "torchsparsegradutils.utils.random_sparse",
-    "torchsparsegradutils.utils.utils",
+    "torchsparsegradutils.utils.convert",
     "torchsparsegradutils.utils.dist_stats_helpers",
     "torchsparsegradutils.encoders.pairwise_encoder",
     "torchsparsegradutils.distributions.sparse_multivariate_normal",
