+++ torchsparsegradutils/__init__.py
@@ -1,8 +1,13 @@
-from .indexed_matmul import gather_mm, segment_mm
-from .sparse_logsumexp import sparse_bidir_logsumexp, sparse_logsumexp
-from .sparse_lstsq import sparse_generic_lstsq
-from .sparse_matmul import sparse_mm
-from .sparse_solve import sparse_generic_solve, sparse_triangular_solve
+from .ops import (
+    gather_mm,
+    segment_mm,
+    sparse_bidir_logsumexp,
+    sparse_generic_lstsq,
+    sparse_generic_solve,
+    sparse_logsumexp,
+    sparse_mm,
+    sparse_triangular_solve,
+)
 
 __all__ = [
     "sparse_mm",
