--- torchsparsegradutils/distributions/constraints.py
@@ -1,1 +1,0 @@
-# TODO: constraints for sparse distributions
