+++ torchsparsegradutils/ops/__init__.py
@@ -1,0 +1,17 @@
+from .generic_solve import sparse_generic_solve
+from .indexed_matmul import gather_mm, segment_mm
+from .logsumexp import sparse_bidir_logsumexp, sparse_logsumexp
+from .lstsq import sparse_generic_lstsq
+from .matmul import sparse_mm
+from .triangular_solve import sparse_triangular_solve
+
+__all__ = [
+    "sparse_mm",
+    "gather_mm",
+    "segment_mm",
+    "sparse_triangular_solve",
+    "sparse_generic_solve",
+    "sparse_generic_lstsq",
+    "sparse_logsumexp",
+    "sparse_bidir_logsumexp",
+]
